requirements.txt
biopython>=1.81
streamlit>=1.25
python-dotenv>=1.0.0
openai>=1.0.0
py3Dmol>=2.0.1
matplotlib>=3.7.0
freesasa>=2.1.0
pandas>=1.5.0
# Add more as needed for visualization or other LLMs 